fix(core): only set zoombar data and sanitize date values if a cartesian time-series chart
